Refactored TransportProtocol tests. Now unix-domain files are temp files, so their name is unique.
